Unify the fixtures
