+++ ndcube/conftest.py
@@ -2,7 +2,6 @@
 This file contains a set of common fixtures to get a set of different but
 predicable NDCube objects.
 """
-
 import astropy.units as u
 import numpy as np
 import pytest
@@ -13,19 +12,13 @@ from ndcube import ExtraCoords, NDCube
 
 
 @pytest.fixture
-def wcs_4d():
+def wcs_4d_t_l_lt_ln():
     header = {
-        'CTYPE4': 'HPLN-TAN',
-        'CUNIT4': 'arcsec',
-        'CDELT4': 5,
-        'CRPIX4': 5,
-        'CRVAL4': 0,
-
-        'CTYPE3': 'HPLT-TAN',
-        'CUNIT3': 'arcsec',
-        'CDELT3': 20,
-        'CRPIX3': 0,
-        'CRVAL3': 0,
+        'CTYPE1': 'TIME    ',
+        'CUNIT1': 'min',
+        'CDELT1': 0.4,
+        'CRPIX1': 0,
+        'CRVAL1': 0,
 
         'CTYPE2': 'WAVE    ',
         'CUNIT2': 'Angstrom',
@@ -33,16 +26,25 @@ def wcs_4d():
         'CRPIX2': 0,
         'CRVAL2': 0,
 
-        'CTYPE1': 'TIME    ',
-        'CUNIT1': 'min',
-        'CDELT1': 0.4,
-        'CRPIX1': 0,
-        'CRVAL1': 0,
-        }
+        'CTYPE3': 'HPLT-TAN',
+        'CUNIT3': 'arcsec',
+        'CDELT3': 20,
+        'CRPIX3': 0,
+        'CRVAL3': 0,
+
+        'CTYPE4': 'HPLN-TAN',
+        'CUNIT4': 'arcsec',
+        'CDELT4': 5,
+        'CRPIX4': 5,
+        'CRVAL4': 0,
+
+        'DATEREF': "2020-01-01T00:00:00"
+    }
     return WCS(header=header)
 
+
 @pytest.fixture
-def wcs_3d():
+def wcs_3d_l_lt_ln():
     header = {
         'CTYPE1': 'WAVE    ',
         'CUNIT1': 'Angstrom',
@@ -61,12 +63,13 @@ def wcs_3d():
         'CDELT3': 10,
         'CRPIX3': 0,
         'CRVAL3': 0,
-        }
+    }
 
     return WCS(header=header)
 
+
 @pytest.fixture
-def wcs_2d_spatial():
+def wcs_2d_lt_ln():
     spatial = {
         'CTYPE1': 'HPLT-TAN',
         'CUNIT1': 'arcsec',
@@ -82,8 +85,9 @@ def wcs_2d_spatial():
     }
     return WCS(header=spatial)
 
+
 @pytest.fixture
-def wcs_1d():
+def wcs_1d_l():
     spatial = {
         'CNAME1': 'spectral',
         'CTYPE1': 'WAVE',
@@ -93,6 +97,56 @@ def wcs_1d():
         'CRVAL1': 0.5,
     }
     return WCS(header=spatial)
+
+
+@pytest.fixture
+def wcs_3d_ln_lt_t_rotated():
+    h_rotated = {
+        'CTYPE1': 'HPLN-TAN',
+        'CUNIT1': 'arcsec',
+        'CDELT1': 0.4,
+        'CRPIX1': 0,
+        'CRVAL1': 0,
+        'NAXIS1': 5,
+
+        'CTYPE2': 'HPLT-TAN',
+        'CUNIT2': 'arcsec',
+        'CDELT2': 0.5,
+        'CRPIX2': 0,
+        'CRVAL2': 0,
+        'NAXIS2': 5,
+
+        'CTYPE3': 'TIME    ',
+        'CUNIT3': 'seconds',
+        'CDELT3': 3,
+        'CRPIX3': 0,
+        'CRVAL3': 0,
+        'NAXIS3': 2,
+
+        'DATEREF': "2020-01-01T00:00:00",
+
+        'PC1_1': 0.714963912964,
+        'PC1_2': -0.699137151241,
+        'PC1_3': 0.0,
+        'PC2_1': 0.699137151241,
+        'PC2_2': 0.714963912964,
+        'PC2_3': 0.0,
+        'PC3_1': 0.0,
+        'PC3_2': 0.0,
+        'PC3_3': 1.0
+    }
+    return WCS(header=h_rotated)
+
+
+@pytest.fixture
+def simple_extra_coords_3d():
+    return ExtraCoords.from_lookup_tables(('time', 'hello', 'bye'),
+                                          (0, 1, 2),
+                                          (list(range(2)) * u.pix,
+                                           list(range(3)) * u.pix,
+                                           list(range(4)) * u.pix
+                                           )
+                                          )
 
 
 def data_nd(shape):
@@ -107,43 +161,43 @@ def generate_time_extra_coord(data_cube):
 
 
 @pytest.fixture
-def ndcube_4d_simple(wcs_4d):
+def ndcube_4d_ln_lt_l_t(wcs_4d_t_l_lt_ln):
     shape = (5, 8, 10, 12)
+    wcs_4d_t_l_lt_ln.array_shape = shape
     data_cube = data_nd(shape)
-    return NDCube(data_cube, wcs=wcs_4d)
+    return NDCube(data_cube, wcs=wcs_4d_t_l_lt_ln)
 
 
 @pytest.fixture
-def ndcube_4d_uncertainty(wcs_4d):
+def ndcube_4d_uncertainty(wcs_4d_t_l_lt_ln):
     shape = (5, 8, 10, 12)
     data_cube = data_nd(shape)
     uncertainty = np.sqrt(data_cube)
-    return NDCube(data_cube, wcs=wcs_4d, uncertainty=uncertainty)
+    return NDCube(data_cube, wcs=wcs_4d_t_l_lt_ln, uncertainty=uncertainty)
 
 
 @pytest.fixture
-def ndcube_4d_mask(wcs_4d):
+def ndcube_4d_mask(wcs_4d_t_l_lt_ln):
     shape = (5, 8, 10, 12)
     data_cube = data_nd(shape)
     uncertainty = np.sqrt(data_cube)
     mask = data_cube % 2
-    return NDCube(data_cube, wcs=wcs_4d, uncertainty=uncertainty, mask=mask)
+    return NDCube(data_cube, wcs=wcs_4d_t_l_lt_ln, uncertainty=uncertainty, mask=mask)
 
 
 @pytest.fixture
-def ndcube_4d_extra_coords(wcs_4d):
+def ndcube_4d_extra_coords(wcs_4d_t_l_lt_ln, simple_extra_coords_3d):
     shape = (5, 8, 10, 12)
     data_cube = data_nd(shape)
-    extra_coords = generate_time_extra_coord(data_cube)
-    return NDCube(data_cube, wcs=wcs_4d, extra_coords=extra_coords)
+    return NDCube(data_cube, wcs=wcs_4d_t_l_lt_ln, extra_coords=simple_extra_coords_3d)
 
 
 @pytest.fixture
-def ndcube_4d_unit_uncertainty(wcs_4d):
+def ndcube_4d_unit_uncertainty(wcs_4d_t_l_lt_ln):
     shape = (5, 8, 10, 12)
     data_cube = data_nd(shape)
     uncertainty = np.sqrt(data_cube)
-    return NDCube(data_cube, wcs=wcs_4d,
+    return NDCube(data_cube, wcs=wcs_4d_t_l_lt_ln,
                   unit=u.J, uncertainty=uncertainty)
 
 
@@ -156,10 +210,39 @@ def ndcube_4d(request):
 
 
 @pytest.fixture
-def ndcube_2d_simple(wcs_2d_spatial):
+def ndcube_3d_ln_lt_l(wcs_3d_l_lt_ln, simple_extra_coords_3d):
+    shape = (2, 3, 4)
+    wcs_3d_l_lt_ln.array_shape = shape
+    data = data_nd(shape)
+    mask = data > 0
+    return NDCube(
+        data,
+        wcs_3d_l_lt_ln,
+        mask=mask,
+        uncertainty=data,
+        extra_coords=simple_extra_coords_3d
+    )
+
+
+@pytest.fixture
+def ndcube_3d_rotated(wcs_3d_ln_lt_t_rotated, simple_extra_coords_3d):
+    data_rotated = np.array([[[1, 2, 3, 4, 6], [2, 4, 5, 3, 1], [0, -1, 2, 4, 2], [3, 5, 1, 2, 0]],
+                             [[2, 4, 5, 1, 3], [1, 5, 2, 2, 4], [2, 3, 4, 0, 5], [0, 1, 2, 3, 4]]])
+    mask_rotated = data_rotated >= 0
+    return NDCube(
+        data_rotated,
+        wcs_3d_ln_lt_t_rotated,
+        mask=mask_rotated,
+        uncertainty=data_rotated,
+        extra_coords=simple_extra_coords_3d
+    )
+
+
+@pytest.fixture
+def ndcube_2d_ln_lt(wcs_2d_lt_ln):
     shape = (10, 12)
     data_cube = data_nd(shape)
-    return NDCube(data_cube, wcs=wcs_2d_spatial)
+    return NDCube(data_cube, wcs=wcs_2d_lt_ln)
 
 
 @pytest.fixture
@@ -171,7 +254,33 @@ def ndcube_2d(request):
 
 
 @pytest.fixture
-def ndcube_1d_simple(wcs_1d):
+def ndcube_1d_l(wcs_1d_l):
     shape = (10,)
     data_cube = data_nd(shape)
-    return NDCube(data_cube, wcs=wcs_1d)
+    return NDCube(data_cube, wcs=wcs_1d_l)
+
+
+@pytest.fixture(params=[
+    "ndcube_4d_ln_lt_l_t",
+    "ndcube_4d_uncertainty",
+    "ndcube_4d_mask",
+    "ndcube_4d_extra_coords",
+    "ndcube_4d_unit_uncertainty",
+    "ndcube_3d_ln_lt_l",
+    "ndcube_3d_rotated",
+    "ndcube_2d_ln_lt",
+    "ndcube_1d_l",
+])
+def all_ndcubes(request):
+    """
+    All the above ndcube fixtures in order.
+    """
+    return request.getfixturevalue(request.param)
+
+
+@pytest.fixture
+def ndc(request):
+    """
+    A fixture for use with indirect to lookup other fixtures.
+    """
+    return request.getfixturevalue(request.param)
